+++ web/src/main/java/cn/v5cn/v5cms/biz/impl/SiteBizImpl.java
@@ -28,7 +28,7 @@ public class SiteBizImpl implements SiteBiz {
     }
 
     @Override
-    public int addSite(Site site) {
+    public Long addSite(Site site) {
         site.setCreateDate(DateTime.now().toDate());
         return siteDao.addSite(site);
     }
